Move modules into their own space
Modules where initially defined in the collector space as the only trait
they were implementing was the collector one. But that will change
sometime in the future with post-processing. Moving them to their own
dedicated space will help their development so the code is located in a
single space.

No implementation change intended, this should be purely cosmetic.

Signed-off-by: Antoine Tenart <atenart@redhat.com>