user: add runtime symbol resolution
USDT symbols have to be resolved at runtime because the information
stored in the elf headers are just offsets while the instruction pointer
we receive from EBPF is a virtual address.

Support virtual to offset address calculation and USDT symbol
resolution. In the userspace section, we parse the binary every time
which is very inefficient. This will be improved in a later patch.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>